requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
motor>=3.3.2
prometheus-client>=0.19.0
colorlog>=6.8.0
pyyaml>=6.0.1
aiohttp>=3.9.1
pytest>=7.4.3
black>=23.12.1 